Fix docker containers and preload schema for tests.
Docker containers for all vendors now restart successfully.
Logged output is less verbose.
The ApacheDS container loads custom schemas.

+++ docker/openldap/entrypoint.sh
@@ -19,22 +19,19 @@
 
 ulimit -n 1024
 
-# if [[ $(/usr/sbin/slaptest) == *"not configured"* ]]; then
-# if no backups exist
-# if [ -z "$(ls -A /var/backups/slapd*)" ]; then
-# if [ ! -f /.done ]; then
+LDAP_LOG_LEVEL=${LDAP_LOG_LEVEL:-256}
+LDAP_BACKEND=${LDAP_BACKEND:-mdb}
+
+if [ ! -d "/var/lib/ldap/data.mdb" ]; then
 
   echo "Configuring OpenLDAP server..."
-
-  LDAP_LOG_LEVEL=${LDAP_LOG_LEVEL:-256}
-  LDAP_BACKEND=${LDAP_BACKEND:-mdb}
 
   chown -R openldap:openldap /var/lib/ldap
   chown -R openldap:openldap /etc/ldap
 
   rm -R /var/backups
 
-cat <<EOF | debconf-set-selections
+  cat <<EOF | debconf-set-selections
 slapd slapd/internal/generated_adminpw password ${LDAP_ADMIN_PASSWORD}
 slapd slapd/internal/adminpw password ${LDAP_ADMIN_PASSWORD}
 slapd slapd/password2 password ${LDAP_ADMIN_PASSWORD}
@@ -51,14 +48,34 @@ EOF
 
   dpkg-reconfigure -f noninteractive slapd
 
-#   if [[ $(/usr/sbin/slaptest) == *"succeeded"* ]]; then
-#     echo "set up complete"
-#     touch /.done
-#   fi
-# fi
+  echo "Testing OpenLDAP server..."
+
+  /usr/sbin/slaptest
+fi
 
 echo "Starting OpenLDAP server..."
 
+# usage: slapd options
+#         -4              IPv4 only
+#         -6              IPv6 only
+#         -T {acl|add|auth|cat|dn|index|passwd|test}
+#                         Run in Tool mode
+#         -c cookie       Sync cookie of consumer
+#         -d level        Debug level
+#         -f filename     Configuration file
+#         -F dir  Configuration directory
+#         -g group        Group (id or name) to run as
+#         -h URLs         List of URLs to serve
+#         -l facility     Syslog facility (default: LOCAL4)
+#         -n serverName   Service name
+#         -o <opt>[=val] generic means to specify options; supported options:
+#                 slp[={on|off|(attrs)}] enable/disable SLP using (attrs)
+#         -r directory    Sandbox directory to chroot to
+#         -s level        Syslog level
+#         -u user         User (id or name) to run as
+#         -V              print version info (-VV exit afterwards, -VVV print
+#                         info about static overlays and backends)
+#
 /usr/sbin/slapd \
   -h "ldap://$HOSTNAME ldaps://$HOSTNAME ldapi:///" \
   -u openldap \
